+++ common/src/main/java/net/mehvahdjukaar/every_compat/api/EveryCompatAPI.java
@@ -4,8 +4,13 @@ import net.mehvahdjukaar.every_compat.EcTempPluginStorage;
 import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.api.example.WoodGoodAddonExample;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
+import java.util.function.Supplier;
+
+import static net.mehvahdjukaar.every_compat.EveryCompat.maybeAddModule;
 import java.util.function.Supplier;
 
 /**
@@ -37,11 +42,48 @@ public class EveryCompatAPI {
         return EveryCompat.getModulesOfMod(modId);
     }
 
-    /// If you mod has compat mods that support it with Biomes O' Plenty or other Wood Mods below can make an exception
-    /// so EC won't generate blocks from your mod with Biomes O' Plenty
-    public static void addOtherCompatMod(String compatModId, List<String> fromModId, List<String> supportedModId){
-        EcTempPluginStorage.addOtherCompatMod(compatModId, fromModId, supportedModId);
+
+    //TODO: delete these 2 they arent needed. less stuff here the better
+    /// Add your module to EveryCompat. The module will be loaded if the mod is installed
+    @Deprecated(forRemoval = true)
+    public static void addIfLoaded(String modId, Supplier<Class<? extends CompatModule>> moduleClassSupplier) {
+        maybeAddModule(modId, moduleClassSupplier);
     }
 
+    //delete
+    @Deprecated(forRemoval = true)
+    @SafeVarargs
+    /// Same as addIfLoaded but Multiple Module can be StoneModule, MudModule, WoodModule or others for the one same mod
+    public static void addMultipleIfLoaded(String modId, Supplier<Class<? extends CompatModule>>... moduleClasses) {
+        for (Supplier<Class<? extends CompatModule>> moduleClassSupplier : moduleClasses) {
+            maybeAddModule(modId, moduleClassSupplier);
+        }
+    }
+
+    /// If you mod has compat mods that support it with Biomes O' Plenty or other Wood Mods below can make an exception
+    /// so EC won't generate blocks from your mod with Biomes O' Plenty
+    public static void addOtherCompatMod(String compatModId, List<String> fromModId, List<String> supportedModId) {
+        EcTempPluginStorage.addCompatMod(compatModId, fromModId, supportedModId);
+    }
+
+    public static void addOtherCompatMod(String compatModId, String fromModId, String supportedModId) {
+        addOtherCompatMod(compatModId, List.of(fromModId), List.of(supportedModId));
+    }
+
+    public static void addOtherCompatMod(String compatModId, String fromModId, String... supportedModId) {
+        List<String> list = new ArrayList<>();
+        Collections.addAll(list, supportedModId);
+        addOtherCompatMod(compatModId, List.of(fromModId), list);
+    }
+
+    public static void addOtherCompatMod(String compatModId, List<String> fromModId, String... supportedModId) {
+        List<String> list = new ArrayList<>();
+        Collections.addAll(list, supportedModId);
+        addOtherCompatMod(compatModId, fromModId, list);
+    }
+
+    public static void addOtherCompatMod(String compatModId, List<String> fromModId, String supportedModId) {
+        addOtherCompatMod(compatModId, fromModId, List.of(supportedModId));
+    }
 
 }