+++ common/src/main/java/net/mehvahdjukaar/every_compat/modules/friendsandfoes/FriendsAndFoesModule.java
@@ -3,7 +3,6 @@ package net.mehvahdjukaar.every_compat.modules.friendsandfoes;
 import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.api.SimpleEntrySet;
 import net.mehvahdjukaar.every_compat.api.SimpleModule;
-import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.mehvahdjukaar.moonlight.api.set.wood.VanillaWoodTypes;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodType;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
@@ -16,13 +15,15 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 
+import static net.mehvahdjukaar.every_compat.common_classes.poiUtility.simpleAddBlocksToPOI;
+
 //SUPPORT: v3.0.9+
 public class FriendsAndFoesModule extends SimpleModule {
 
     public final SimpleEntrySet<WoodType, Block> beehives;
 
     public FriendsAndFoesModule(String modId) {
-        super(modId, "faf");
+        super(modId, "faf", EveryCompat.MOD_ID);
 
         beehives = SimpleEntrySet.builder(WoodType.class, "beehive",
                         getModBlock("spruce_beehive"), () -> VanillaWoodTypes.SPRUCE,
@@ -46,11 +47,7 @@ public class FriendsAndFoesModule extends SimpleModule {
     @Override
     public void onModInit() {
         super.onModInit();
-        RegHelper.addExtraPOIStatesRegistration(event ->
-                beehives.blocks.values().forEach(b ->
-                        event.addBlockToPoi(PoiTypes.BEEHIVE, b)
-                ));
-
+        simpleAddBlocksToPOI(beehives, PoiTypes.BEEHIVE);
     }
 
 }